Apache Kafka Mini Project/detector/requirements.txt
kafka-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib